+++ adalflow/CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.2.4] - 2024-09-23
+
+### Improved
+- Better diagnose report for `Trainer.diagnose`.
+- Multi-hop RAG with handling of Cycle.
 ## [0.2.3] - 2024-09-20
 
 ### Rename